bigint updated

+++ org_init_scripts/072_pos_counter_transaction_breakup.sql
@@ -1,14 +1,14 @@
 create table if not exists pos_counter_transaction_breakup
 (
-    voucher_id     int   not null,
-    account_id     int   not null,
-    account_name   text  not null,
-    pos_counter_id int   not null,
-    credit         float not null default 0,
-    debit          float not null default 0,
-    amount         float not null generated always as (debit - credit) stored,
-    session_id     int,
-    settlement_id  int,
+    voucher_id     bigint not null,
+    account_id     bigint not null,
+    account_name   text   not null,
+    pos_counter_id bigint not null,
+    credit         float  not null default 0,
+    debit          float  not null default 0,
+    amount         float  not null generated always as (debit - credit) stored,
+    session_id     bigint,
+    settlement_id  bigint,
     primary key (voucher_id, account_id)
 );
 --##
@@ -41,7 +41,7 @@ create function update_pos_counter_txn(voucher, json)
     returns bool as
 $$
 declare
-    v_session_id int;
+    v_session_id bigint;
     acc          account;
     item         pos_counter_transaction_breakup;
     items        pos_counter_transaction_breakup[] := (select array_agg(x)
